requirements.txt
asgiref==3.3.1
cryptography==3.4.4
Django==3.1.4
django-allauth==0.44.0
django-phonenumber-field==5.0.0
django-simple-bulma==2.1.0
django-upload-validator==1.1.5
libsass==0.20.1
oauthlib==3.1.0
phonenumberslite==8.12.17
Pillow==8.1.0
PyJWT==2.0.1
python-magic==0.4.18
python3-openid==3.2.0
pytz==2020.4
requests-oauthlib==1.3.0
sqlparse==0.4.1
